+++ loopx/domain_packs/__init__.py
@@ -1,0 +1,1 @@
+"""Domain capability packs for LoopX."""
